requirements.txt
fastapi>=0.111.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
sqlalchemy==2.0.27
pydantic==2.7.3
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest-playwright
fastapi-users[sqlalchemy]==12.1.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
openai
httpx==0.27.0
celery==5.3.6
redis==5.0.3
psycopg[binary]==3.1.18
python-jose[cryptography]==3.3.0
tiktoken
yfinance
humanize